# KAFKA-5746: Add new metrics to support health checks (KIP-188)