--- a/clients/src/main/java/org/apache/kafka/common/metrics/Gauge.java
+++ b/clients/src/main/java/org/apache/kafka/common/metrics/Gauge.java
@@ -23,8 +23,15 @@ package org.apache.kafka.common.metrics;
 public interface Gauge<T> {
 
     /**
-     * Returns the current value associated with this gauge
+     * Returns the current value associated with this gauge.
+     * <p>
+     * In the future we may introduce a super-interface that is implemented
+     * by {@link Measurable} as well with a default implementation that returns
+     * {@link Measurable#measure(MetricConfig, long)} (for Java8 and above).
+     * </p>
+     * @param config The configuration for this metric
+     * @param now The POSIX time in milliseconds the measurement is being taken
      */
-    public T value();
+    public T value(MetricConfig config, long now);
 
 }