--- a/clients/src/main/java/org/apache/kafka/common/metrics/Measurable.java
+++ b/clients/src/main/java/org/apache/kafka/common/metrics/Measurable.java
@@ -19,7 +19,7 @@ package org.apache.kafka.common.metrics;
 /**
  * A measurable quantity that can be registered as a metric
  */
-public interface Measurable extends MetricValue<Double> {
+public interface Measurable extends MetricValueProvider<Double> {
 
     /**
      * Measure this quantity and return the result as a double