@ijuma 2017-09-21
Would it make sense to merge the SSL authentication PR first?