@rajinisivaram 2017-09-22
@ijuma Thank you for the review, I have addressed the comments. But yes, it will be better to merge the SSL authentication PR first.