@rajinisivaram 2017-09-26
@ijuma Added metrics for `Errors.NONE`