@rajinisivaram 2017-09-27
Preliminary performance test results, running producer performance on my laptop:
This does not show any noticeable difference beyond the variation in runs, but note that this doesn't reflect the overheads for fetch requests with small wait and large number of partitions where the difference may be more significant:

**Maximum throughput (100 byte records, no throughput limit):**
```
Trunk:
50000000 records sent, 977822.974929 records/sec (93.25 MB/sec), 90.57 ms avg latency, 3187.00 ms max latency, 3 ms 50th, 213 ms 95th, 3032 ms 99th, 3158 ms 99.9th.
50000000 records sent, 1001021.041462 records/sec (95.46 MB/sec), 116.89 ms avg latency, 4319.00 ms max latency, 2 ms 50th, 183 ms 95th, 4123 ms 99th, 4293 ms 99.9th.
50000000 records sent, 1000500.250125 records/sec (95.42 MB/sec), 117.71 ms avg latency, 5012.00 ms max latency, 2 ms 50th, 186 ms 95th, 4167 ms 99th, 4972 ms 99.9th.
With this PR:
50000000 records sent, 998043.834085 records/sec (95.18 MB/sec), 113.76 ms avg latency, 4705.00 ms max latency, 2 ms 50th, 194 ms 95th, 4009 ms 99th, 4640 ms 99.9th.
50000000 records sent, 1000640.409862 records/sec (95.43 MB/sec), 107.75 ms avg latency, 4016.00 ms max latency, 2 ms 50th, 189 ms 95th, 3896 ms 99th, 4001 ms 99.9th.
50000000 records sent, 999520.230289 records/sec (95.32 MB/sec), 109.99 ms avg latency, 5029.00 ms max latency, 2 ms 50th, 157 ms 95th, 3912 ms 99th, 4996 ms 99.9th.
```
Latency with 950K records/second
```
Trunk:
50000000 records sent, 899879.416158 records/sec (85.82 MB/sec), 2.25 ms avg latency, 111.00 ms max latency, 2 ms 50th, 3 ms 95th, 6 ms 99th, 81 ms 99.9th
50000000 records sent, 899895.612109 records/sec (85.82 MB/sec), 2.46 ms avg latency, 153.00 ms max latency, 2 ms 50th, 5 ms 95th, 15 ms 99th, 61 ms 99.9th.
50000000 records sent, 899879.416158 records/sec (85.82 MB/sec), 2.46 ms avg latency, 148.00 ms max latency, 2 ms 50th, 3 ms 95th, 9 ms 99th, 143 ms 99.9th.
With this PR
50000000 records sent, 899830.831804 records/sec (85.81 MB/sec), 2.18 ms avg latency, 113.00 ms max latency, 2 ms 50th, 3 ms 95th, 6 ms 99th, 80 ms 99.9th.
50000000 records sent, 899863.220790 records/sec (85.82 MB/sec), 2.21 ms avg latency, 107.00 ms max latency, 2 ms 50th, 3 ms 95th, 6 ms 99th, 80 ms 99.9th.
50000000 records sent, 899830.831804 records/sec (85.81 MB/sec), 2.40 ms avg latency, 103.00 ms max latency, 2 ms 50th, 4 ms 95th, 9 ms 99th, 79 ms 99.9th.
```